update setup data dir script

+++ setup-data-dir.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-import os, sys, re
+import os
 import argparse
 from lib.utils import test, make_symlink, ls, ObjAsDict, load_config_py, sysexecOut
 
@@ -32,7 +32,6 @@ def setup_data_dir(newdir=None):
         load_config_py(info_file, ObjAsDict(Settings))
 
     user = os.environ["USER"].strip() if "USER" in os.environ else sysexecOut("whoami").strip() # logname for mac with zsh
-    print(user)
     assert user
     workdir = "%s/%s/setups-data/%s" % (Settings.workdir_base, user, curdir[len(setup_dir_prefix):])
     print("Work dir: %s" % workdir)
@@ -41,17 +40,12 @@ def setup_data_dir(newdir=None):
         print("Work dir does not exist, create...")
         os.makedirs(workdir)
 
-    if not os.path.islink(workdir + "/setup-dir-symlink"):
-        print("Workdir/setup-dir-symlink does not exist, create...")
-        os.symlink(curdir, workdir + "/setup-dir-symlink")
-    #else:
-    #	assert os.readlink(workdir + "/setup-dir-symlink") == curdir
-
     if os.path.islink(datadirlink) and os.readlink(datadirlink) != workdir:
         print("existing missmatching symlink:")
         ls(datadirlink)
         print("Deleting %s..." % datadirlink)
         os.remove(datadirlink)
+
     if not os.path.islink(datadirlink):
         print("%s does not exist, create..." % datadirlink)
         os.symlink(workdir, datadirlink)
@@ -69,7 +63,6 @@ def setup_data_dir(newdir=None):
     make_symlink(os.path.join(datadirlink, newdir), newdir)
     ls(newdir)
 
-    print("Finished.")
 
 
 def auto_update_dirs():
@@ -88,27 +81,17 @@ def auto_update_dirs():
 def setup_dataset_dir():
     """Symlink to the dataset"""
     curdir = os.getcwd()
-    print("Current dir: %s" % curdir)
-    setup_dir_prefix = os.path.realpath(os.path.expanduser("~/setups")) + "/"
-    test(curdir.startswith(setup_dir_prefix), "Error: Must be in %s" % setup_dir_prefix)
-    dir_info = curdir + "/dataset-dir-info.py"
-    if not os.path.exists(dir_info):
-        print("Couldnt find %s!" % dir_info, "Please make sure you are calling\
-              this function from the base file and the info file exists.")
-        return
-
-    # Load the path from dataset-dir-info.py
-    load_config_py(dir_info, ObjAsDict(Settings))
     if not os.path.exists(Settings.dataset): 
-        print(""" WARN: The specified path to dataset %s doesn't exist. Please put the
- correct path in %s and run again if you want a symlink to the dataset!""" % (Settings.dataset ,dir_info))
+        print(""" WARN: The specified path to dataset %s doesn't exist. Please
+              put the correct path in %s and run again if you want a symlink to
+              the dataset!""" % (Settings.dataset, find_info_file(setup_dir_prefix[:-1], curdir)))
     else:
         # Create the Symlink
-        print("Creating dataset_symlink.")
         dataset_symlink = curdir + "/dataset_symlink"
         if os.path.exists(dataset_symlink):
             print("Removing old dataset_symlink")
             os.remove(dataset_symlink)
+        print("Creating dataset_symlink.")
         os.symlink(Settings.dataset, dataset_symlink)
 
 
@@ -116,12 +99,16 @@ def setup_pkg_dir():
     """ Symlink to github packages in ~/return/pkg."""
     curdir = os.getcwd()
     pkg_path = os.path.join(os.path.expanduser("~"), "returnn/pkg")
-    print("Creating pkg_symlink.")
-    pkg_symlink = curdir + "/pkg_symlink"
-    if os.path.exists(pkg_symlink):
-        print("Removing old pkg_symlink")
-        os.remove(pkg_symlink)
-    os.symlink(pkg_path, pkg_symlink)
+    if not os.path.exists(pkg_path): 
+        print("WARN: The specified path to dataset %s doesn't exist." % pkg_path)
+    else:
+        # Create the Symlink
+        pkg_symlink = curdir + "/pkg_symlink"
+        if os.path.exists(pkg_symlink):
+            print("Removing old pkg_symlink")
+            os.remove(pkg_symlink)
+        print("Creating pkg_symlink.")
+        os.symlink(pkg_path, pkg_symlink)
 
 
 if __name__ == "__main__":
@@ -141,3 +128,4 @@ if __name__ == "__main__":
             setup_data_dir(d)
     setup_dataset_dir()
     setup_pkg_dir()
+    print("Finished.")
